requirements.txt
numpy
pyvista
pyvistaqt
ikpy
xarm-python-sdk
pyserial